src/PillowEg/pyproject.toml
[project]
name = "pillowprojects"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "nuitka>=2.6.7",
    "pillow>=11.1.0",
]
